refactor: Rename pdbml_converter module to serializers for concise naming
Rename sloth/pdbml_converter.py to sloth/serializers.py to adopt a more
concise, one-word module name that better aligns with modern Python
conventions for data transformation and output formatting.

### 🔧 Core Changes:

**File Operations:**
- Renamed sloth/pdbml_converter.py → sloth/serializers.py (file rename, no content changes)
- Updated import statements across 7 files to use serializers module

**Updated Import Statements in:**
- sloth/__init__.py: Main package exports
- tests/test_imports.py: Converter class imports and module structure tests
- tests/test_pdbml_conversion.py: Pipeline and dictionary parser imports
- tests/test_permissive_mode.py: Pipeline import
- tests/test_pipeline_integration.py: Converter and mapping generator imports
- tests/test_relationship_resolution.py: Multiple serializer class imports

**Classes Relocated to serializers.py:**
- PDBMLConverter: Core mmCIF → PDBML XML conversion
- XMLMappingGenerator: Schema and mapping rule generation
- DictionaryParser: mmCIF dictionary parsing logic
- RelationshipResolver: Parent-child relationship resolution
- MMCIFToPDBMLPipeline: Complete conversion pipeline

### 📂 Files Modified: 8 files
- 1 file renamed: pdbml_converter.py → serializers.py
- 7 files updated: Import statement changes across core and test modules

### 🎯 Benefits:

1. **Concise Naming**: One-word module name follows Python conventions
2. **Modern Convention**: "Serializers" is widely recognized for data transformation
3. **Broader Scope**: Name encompasses XML generation, relationship handling, and data export
4. **Framework Alignment**: Consistent with popular frameworks (Django, DRF) naming patterns

### ✅ Quality Assurance:

- Pure rename operation - no functional changes
- All imports updated consistently across codebase
- Comprehensive test coverage maintained
- Full backwards compatibility through proper import updates

This refactoring modernizes the module naming while maintaining all existing
functionality and improving code organization
